+++ module/actor/actor.js
@@ -68,6 +68,17 @@ export class Actor4e extends Actor {
 		this._displayScrollingDamage(options.dhp);
 	}
 
+
+	/* --------------------------------------------- */
+
+	/** @override */
+	applyActiveEffects() {
+		// The Active Effects do not have access to their parent at preparation time so we wait until this stage to
+		// determine whether they are suppressed or not.
+		this.effects.forEach(e => e.determineSuppression());
+		return super.applyActiveEffects();
+	}
+
 	/* -------------------------------------------- */
 
 	/**